apps/platform/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
python-dotenv==1.0.1
alembic==1.13.2
-e /app/packages/shared
